chore: Move files and structure a bit
