+++ flottform/demo/src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { base } from '$app/paths';
 	import '../app.postcss';
-	import '@flottform/forms/theme/base.css';
+	import '@flottform/forms/base.css';
 	import '@flottform/forms/theme/default.css';
 </script>
