+++ flottform/forms/src/index.ts
@@ -3,4 +3,4 @@ export { FlottformFileInputClient } from './flottform-file-input-client.ts';
 export { FlottformTextInputClient } from './flottform-text-input-client.ts';
 export { FlottformTextInputHost } from './flottform-text-input-host.ts';
 export type { ClientState, SafeEndpointInfo, Logger } from './internal.ts';
-export { createDefaultFlottformComponent } from './theme/default-component.ts';
+export { createDefaultFlottformComponent } from './default-component.ts';
